Refactor code to separate reading and writing into two passes
This commit doesn't change functionality, but makes the code cleaner.
Previously, the code would traverse the source directory and render
each file encountered to the destination directory in a single pass.
This commit makes the program read all the data to an in-memory
data structure, then write the in-memory data structure to the
destination directory.